Fixing negative numbers in s-expression parser.
The check to see if the first byte is 0..9 was added for performance (to
avoid trying to parse every literal as a number) but - was forgotten.  I
could have added other literals like + there as well, but I think - is
the only one that's actually required for non-"weird" floats.