Resolve failing 64-bit configurations for Appveyor
After a lot of poking around using RDP to investigate 64-bit build
configurations that appeared to freeze, it seems like incompatible OpenSSL
DLLs were the culprit.

The freezing was due to a rather cryptic error dialog that showed itself
only when the `lib` test executable is run. The failing builds all
appeared to be loading OpenSSL DLLs that were not compatible with the
build target. Once OpenSSL DLLs were copied into the same directory as the
test executable, the error was gone, likely due to how standard DLL search
prefers DLLs in the same directory as the program executing.